Cleaned a few things up.
