+++ my-scheme-interpreter/src/lib.rs
@@ -1,12 +1,10 @@
 pub mod array;
-pub mod cell_mut;
+pub mod cell;
 pub mod list_procs;
 //pub mod builtin;
-//pub mod cell;
 pub mod char;
 //pub mod core_proc;
 pub mod env;
-//pub mod env_impl;
 pub mod err;
 //pub mod error;
 //pub mod eval_helpers;
@@ -20,5 +18,4 @@ pub mod value;
 //pub mod scheme_libs;
 pub mod string;
 pub mod types;
-//pub mod vector;
 //pub mod vm;
